ui_frontend/requirements.txt
streamlit==1.37.1
pillow>=10.3.0
Requests>=2.31.0

